requirements.txt
Django==2.2.16
djangorestframework
django-rest-swagger
drf-yasg

